// interface/search/search-applicabilities-panel.interface.ts
import { SearchApplicabilitiesInterface } from '~/interface/search/data/search-applicabilities.interface'

export interface SearchApplicabilitiesPanelInterface {
  selectedMarka: number[] | []
  selectedModel: number[] | []
  selectedGenerations: number[] | []
  dataModel: SearchApplicabilitiesInterface[] | []
  dataGenerations: SearchApplicabilitiesInterface[] | []
}
export interface SearchApplicabilitiesPanelStore {
  searchApplicabilitiesPanel: SearchApplicabilitiesPanelInterface[]
}